turn into a visual studio project
