+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\NeuralNet"
+  ],
+  "SelectedNode": "\\NeuralNet\\NeuralNet.py",
+  "PreviewInSolutionExplorer": false
+}
