+++ README.md
@@ -1,3 +1,3 @@
-# Neural-Network
-Contains all of the current models of my attempts at implementing a Neural Network that incorporates parts of Stochastic Gradient Descent.
-None of the programs currently work in their present form.
+#Neural Network
+Contains all of the current models of my attempts at implementing a Neural Network that incorporates parts of Stochastic Gradient Descent. None of the programs currently work in their present form.
+
